Riclassificazione del ruleParser come oggetto a se stante presente nella classe Main e non piu presente nella classe Unita Immobiliare, inclusione delle funzioni di ruleProcessing all'interno della classe RuleParser
