+++ src/inge/progetto/Main.java
@@ -38,6 +38,7 @@ public class Main {
         ArrayList<ModalitaOperativa> listaModalitaOperative = new ArrayList<>();
         ArrayList<CategoriaAttuatore> listaCategoriaAttuatori = new ArrayList<>();
         ArrayList<CategoriaSensore> listaCategoriaSensori = new ArrayList<>();
+        RuleParser ruleParser = new RuleParser();
         int contatoreOperazioni = 0;
     /*
         CategoriaAttuatore cateAtt1 = new CategoriaAttuatore("cateAtt1", "testo");
@@ -167,7 +168,6 @@ public class Main {
                                 if (!giaPresente) {
 
                                     UnitaImmobiliare temp = new UnitaImmobiliare(tipo, nome);
-
                                     listaUnitaImmobiliari.add(temp);
                                     System.out.println("\n*** Unità immobiliare creata correttamente *** ");
                                     unitaImmobiliare = temp;
@@ -1059,8 +1059,7 @@ public class Main {
 
                 do {
                     if (contatoreOperazioni == 4 && !unitaImmobiliare.getTipo().equals("")) {
-                        String readRules = unitaImmobiliare.getRegole().readRuleFromFile();
-                        applyRules(unitaImmobiliare.getListaSensori(), unitaImmobiliare.getListaAttuatori(), readRules);
+                        ruleParser.applyRules(unitaImmobiliare.getListaSensori(),unitaImmobiliare.getListaAttuatori());
                         contatoreOperazioni = 0;
                         unitaImmobiliare.refreshLetture();
 
@@ -1097,6 +1096,7 @@ public class Main {
                                         if (immo.getNome().equals(seleziona)) {
                                             presente = true;
                                             unitaImmobiliare = immo;
+                                            ruleParser.setFileName(seleziona+".txt");
                                             break;
                                         }
                                     }
@@ -1215,11 +1215,9 @@ public class Main {
                             int sceltaVisualizza;
                             do {
                                 if (contatoreOperazioni == 4 && !unitaImmobiliare.getTipo().equals("")) {
-                                    String readRules = unitaImmobiliare.getRegole().readRuleFromFile();
-                                    applyRules(unitaImmobiliare.getListaSensori(), unitaImmobiliare.getListaAttuatori(), readRules);
+                                    ruleParser.applyRules(unitaImmobiliare.getListaSensori(),unitaImmobiliare.getListaAttuatori());
                                     contatoreOperazioni = 0;
                                     unitaImmobiliare.refreshLetture();
-
                                 }
                                 contatoreOperazioni++;
 
@@ -1330,7 +1328,7 @@ public class Main {
                             }
 
                             System.out.println("\n...REGOLE ATTUALMENTE CREATE PER QUESTA UNITA' IMMOBILIARE...");
-                            String readRules = unitaImmobiliare.getRegole().readRuleFromFile();
+                            String readRules = ruleParser.readRuleFromFile();
 
                             if (readRules.equals(""))
                                 System.out.println("XX Non sono state definite regole per l'unita immobiliare selezionata XX\n");
@@ -1594,7 +1592,7 @@ public class Main {
                             } while (azione);
 
                             System.out.println("+++ Regola inserita: " + regola);
-                            unitaImmobiliare.getRegole().writeRuleToFile(regola);
+                            ruleParser.writeRuleToFile(regola);
 
                             break;
                         case 0:
@@ -1644,190 +1642,4 @@ public class Main {
     }
 
     //TODO: Aggiungere le funzioni o nella classe UnitaImmob o RuleParser
-    private static void applyRules(ArrayList<Sensore> listaSensori, ArrayList<Attuatore> listaAttuatori, String readRules) {
-        System.out.println("\n\n...APPLICAZIONE REGOLE...\n");
-
-        if (readRules.equals("")) {
-            System.out.println("XX Non sono state inserite regole per l'unita immobiliare XX\n");
-            return;
-        }
-
-        String[] rules = readRules.split("\n");
-
-        for (String r : rules) {
-
-            String r2 = r.replace("IF ", "");
-            String[] tokens = r2.split(" THEN ");
-
-            boolean ris = calculate(tokens[0], listaSensori);
-            System.out.println(r + " :: " + ris);
-            if (ris)
-                applyActions(tokens[1], listaAttuatori);
-        }
-    }
-
-    private static void applyActions(String token, ArrayList<Attuatore> listaAttuatori) {
-        for (String tok : token.split(" ; "))
-            apply(tok, listaAttuatori);
-    }
-
-    private static void apply(String act, ArrayList<Attuatore> listaAttuatori) {
-        String[] toks = act.split(" := ");
-
-        Attuatore actD = null;
-        for (Attuatore att : listaAttuatori) {
-            if (att.getNome().equals(toks[0])) {
-                actD = att;
-            }
-        }
-
-        if (actD == null)
-            return; //siccome abbiamo regole fatte di versioni precedenti...va bene cosi..gestito dal input
-
-        String modPrecedente = actD.getNome() + ": " + actD.getModalitaAttuale() + " -> ";
-        if (toks[1].contains("|")) {
-            String[] againTokens = toks[1].split("\\|");
-
-            if (!againTokens[2].matches("-?[0-9]+"))
-                return;
-
-            actD.setModalitaAttuale(againTokens[0], againTokens[1], Integer.parseInt(againTokens[2]));
-        } else {
-            actD.setModalitaAttuale(toks[1]);
-        }
-        System.out.println(modPrecedente + actD.getModalitaAttuale() + "\n");
-
-    }
-
-    //IL CALCOLO DELLE REGOLE HA UN'ASSOCIAZIONE A DESTRA, oddddioooo, precendenza all'AND!
-    private static boolean calculate(String cos, ArrayList<Sensore> listaSensori) {
-        if (cos.equals("true"))
-            return true;
-
-        if (cos.contains("OR")) {
-            String[] expTok = cos.split(" OR ", 2);
-            return calculate(expTok[0], listaSensori) || calculate(expTok[1], listaSensori);
-        }
-
-        if (cos.contains("AND")) {
-            String[] expTok = cos.split(" AND ", 2);
-            return calculate(expTok[0], listaSensori) && calculate(expTok[1], listaSensori);
-        }
-
-        if (cos.matches("[^<>=\t\n ]+ ([<>=]|<=|>=) [^<>=\t\n ]+")) {
-            String[] expTok = cos.split(" ");
-            return getValExp(expTok, listaSensori);
-
-        }
-        return false;
-
-    }
-
-    private static boolean getValExp(String[] toks, ArrayList<Sensore> listaSensori) {
-        String var1 = toks[0];
-        String operator = toks[1];
-        String var2 = toks[2];
-
-
-        String[] sensVar = var1.split("\\.");
-        Sensore sens1 = null;
-
-        for (Sensore s : listaSensori) {
-            if (s.getNome().equals(sensVar[0])) {
-                sens1 = s;
-                //System.out.println("Sensore s1 trovato");
-                break;
-            }
-        }
-
-        if (sens1 == null)
-            return false;
-
-        Informazione misura1 = sens1.getInformazione(sensVar[1]);
-
-        if (misura1 == null) {
-            System.out.println("Misura s1 non trovata");
-            return false;
-        }
-
-        if (var2.matches("-?[0-9]+")) {
-
-            if (misura1.getTipo().equals("NN"))
-                return false;
-
-            int value = (int) misura1.getValore();
-            int num = Integer.parseInt(var2);
-
-            System.out.println();
-
-            return evalOp(operator, value, num);
-        }
-
-        if (var2.matches("[A-Za-z]([a-zA-Z]|[0-9])*_[A-Za-z]([a-zA-Z]|[0-9])+\\.([a-zA-Z]|[0-9])+")) {
-            String[] sensVar2 = var2.split("\\.");
-            Sensore sens2 = null;
-
-            for (Sensore s : listaSensori) {
-                if (s.getNome().equals(sensVar2[0])) {
-                    sens2 = s;
-                    break;
-                }
-            }
-
-            if (sens2 == null)
-                return false;
-
-            Informazione misura2 = sens2.getInformazione(sensVar2[1]);
-
-            if (misura2 == null)
-                return false;
-
-            if (!misura1.getTipo().equals(misura2.getTipo()))
-                return false;
-
-            if (misura1.getTipo().equals("NN")) {
-                String sca1 = (String) misura1.getValore();
-                String sca2 = (String) misura2.getValore();
-
-                if (!operator.equals("="))
-                    return false;
-                else
-                    return sca1.equals(sca2);
-
-            } else {
-                int value1 = (int) misura1.getValore();
-                int value2 = (int) misura2.getValore();
-
-                return evalOp(operator, value1, value2);
-            }
-        }
-
-        if (var2.matches("[a-zA-Z]+") && operator.equals("=")) {
-            if (misura1.getTipo().equals("NN")) {
-                String sca1 = (String) misura1.getValore();
-                return var2.equals(sca1);
-            }
-        }
-
-        return false;
-    }
-
-    private static boolean evalOp(String operator, int value1, int value2) {
-        if (operator.startsWith("<")) {
-            if (operator.endsWith("="))
-                return value1 <= value2;
-            else
-                return value1 < value2;
-
-        } else if (operator.startsWith(">")) {
-            if (operator.endsWith("="))
-                return value1 >= value2;
-            else
-                return value1 > value2;
-
-        } else {
-            return value1 == value2;
-
-        }
-    }
 }
